# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Glowish API
  version: "1.0.0"
  description: |
    REST API for the Glowish retail management platform.

    ## Authentication
    - **Staff routes** (`/api/users`, `/api/products`, `/api/orders`, etc.): Bearer JWT via NextAuth session cookie.
    - **Customer routes** (`/api/account/*`): Session cookie from marketplace login.
    - **Public routes** (`/api/marketplace/*`, `/api/health`): No authentication required.

    ## Response envelope
    All endpoints return a consistent JSON envelope:
    ```json
    { "success": true, "data": <payload>, "message": "...", "meta": { "page": 1, "limit": 20, "total": 100 } }
    { "success": false, "error": "Human-readable error message" }
    ```
servers:
  - url: https://your-store.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/auth/2fa/setup:
    post:
      tags: [Auth]
      summary: Initiate 2FA setup (admin/org-admin only)
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Returns TOTP secret and QR code data URL
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: NextAuth v5 session cookie
...